exercicios 1 ao 16

+++ sequencial/exercicio 5.py
@@ -1,3 +1,3 @@
-valor = int(input("Digite o valor em metros:"))
+valor = float(input("Digite o valor em metros:"))
 resultado = valor * 100
 print("o valor",valor, "metros em centimetros é:",resultado)
